+++ frame/society/src/mock.rs
@@ -167,7 +167,7 @@ impl EnvBuilder {
 }
 
 /// Run until a particular block.
-pub fn run_to_block(n: u32) {
+pub fn run_to_block(n: u64) {
 	while System::block_number() < n {
 		if System::block_number() > 1 {
 			System::on_finalize(System::block_number());
